// apps/client_dev/types/dto.d.ts
namespace DataTransfer {
  interface BaseDTO {
    id: string;
  }
  export interface AirportDTO extends BaseDTO {
    name: string;
    iataCode: string;
    icaoCode: string;
    countryCode: string;
    type: Enums.AirportType;
    runwayIds: List<string>;
    vehiclesPresentIds: List<string>;
    planesPresentIds: List<string>;
    routesOriginatingFromAirportIds: List<string>;
    routesDestinedForAirportIds: List<string>;
    crewMembersPresentIds: List<string>;
  }
  export interface AlgorithmResultDTO extends BaseDTO {
    flightId: string;
    planeId: string;
    crewMemberIds: List<string>;
    takeoffRunwayId: string;
    landingRunwayId: string;
    originAirportGroundVehicleIds: List<string>;
    destinationAirportGroundVehicleIds: List<string>;
  }
  export interface AlgorithmRunDTO extends BaseDTO {
    algorithmName: string;
    startTime: Date;
    endTime: Date;
    runtimeInMilliseconds: number;
    parametersJson: string;
    resourcesJson: string;
    constrainsMet: Map<string, boolean>;
    logs: string[];
    isSuccessful: boolean;
    failureReason: string;
    isResultsSaved: boolean;
    resultId: string;
  }
  export interface CertificationDTO extends BaseDTO {
    name: string;
    certificationNumber: string;
    issuer: Enums.CertificationIssuer;
    issuingCountry: Enums.CertificationIssuingCountry;
    expirationDate: Date;
    validityPeriod: number;
    assignableRole: Enums.CrewRoles;
    description: string;
    assignedCrewMember: string;
  }
  export interface CrewDTO extends BaseDTO {
    fullName: string;
    email: string;
    phoneNumber: number;
    role: Enums.CrewRoles;
    certificationIds: List<string>;
    totalFlightHours: number;
    baseAirportId: string;
    availability: CrewAvailability;
  }
  export interface FlightDTO extends BaseDTO {
    passengerCount: number;
    flightRouteId: string;
  }
  export interface PlaneDTO extends BaseDTO {
    model: string;
    registrationNumber: string;
    passengerCapacity: number;
    fuelEfficiency: number;
    maxFlightRange: number;
    lastMaintenance: Date;
    totalFlightHours: number;
    maxTakeoffWeight: number;
    shortestRunwayLengthRequired: number;
    shortestRunwayWidthRequired: number;
    planeStatus: Enums.PlaneAvailability;
    currentLocationId: string;
    aircraftOperator: string;
  }
  export interface RouteDTO extends BaseDTO {
    originAirportId: string;
    destinationAirportId: string;
    distance: number;
    estimatedTime: number;
  }
  export interface RunwayDTO extends BaseDTO {
    length: number;
    width: number;
    surfaceType: Enums.RunwaySurfaceType;
    maxWeightCapacity: number;
    orientation: string;
    airportId: string;
  }
  export interface VehicleDTO extends BaseDTO {
    type: Enums.GroundVehicleType;
    vehicleCode: string;
    capacity: number;
    availability: Enums.GroundVehicleAvailability;
    maintenanceDue: Date;
    airportId: string;
  }
}
export default DataTransfer;